+++ src/Controller/VitrineTuxController.php
@@ -22,9 +22,25 @@ class VitrineTuxController extends AbstractController
     #[IsGranted('ROLE_USER')]
     #[Route('/', name: 'app_vitrine_tux_index', methods: ['GET'])]
     public function index(VitrineTuxRepository $vitrineTuxRepository): Response
-    {
+    {   
+        $privateVitrinesTux = array();
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $privateVitrinesTux = $vitrineTuxRepository->findAll();
+        }
+        else {$user = $this->getUser();
+            if($user) {
+                    $membre = $user->getMembreTux();
+            $privateVitrinesTux = $vitrineTuxRepository->findBy(
+                    [
+                          'ispublic' => false,
+                          'membretux' => $membre
+                    ]);
+            }}
+
         return $this->render('vitrine_tux/index.html.twig', [
-            'vitrine_tuxes' => $vitrineTuxRepository->findAll(),
+            'vitrine_tuxes' => $vitrineTuxRepository->findBy(['ispublic'=>true]),
+            'priv_vitrines_tux' => $privateVitrinesTux,
+            'user' => $this->getUser(),
         ]);
     }
 
@@ -55,8 +71,24 @@ class VitrineTuxController extends AbstractController
     #[Route('/{id}', name: 'app_vitrine_tux_show', requirements: ['id' => '\d+'], methods: ['GET'])]
     public function showAction(VitrineTux $vitrine): Response
     {
-        return $this->render('vitrine_tux/show.html.twig', [
-            'vitrine' => $vitrine,
+        $hasAccess = false;
+        if($this->isGranted('ROLE_ADMIN') || $vitrine->isIspublic()) {
+                $hasAccess = true;
+        }
+        else {
+                $user = $this->getUser();
+                if( $user ) {
+                        $member = $user->getMembreTux();
+                        if ( $member &&  ($member == $vitrine->getMembretux()) ) {
+                                $hasAccess = true;
+                        }
+                }
+        }
+        if(! $hasAccess) {
+                throw $this->createAccessDeniedException("You cannot access the requested resource!");
+        }
+        return $this->render('vitrine/show.html.twig', [
+                'vitrine' => $vitrine,
         ]);
     }
     #[IsGranted('ROLE_USER')]
@@ -69,22 +101,41 @@ class VitrineTuxController extends AbstractController
     ): Response
     {   
         if(! $vitrine->getCartesTux()->contains($carte)) {
-            throw $this->createNotFoundException("Carte introuvable dans cette vitrine.");
+            throw $this->createNotFoundException("Couldn't find such a carte in this vitrine!");
     }
 
-        if(! $vitrine->isIspublic()) {
-            throw $this->createAccessDeniedException("Cette vitrine n'est pas publiée.");
+    $hasAccess = false;
+    if($this->isGranted('ROLE_ADMIN') || $vitrine->isIspublic()) {
+            $hasAccess = true;
+    }
+    else {
+            $user = $this->getUser();
+              if( $user ) {
+                      $member = $user->getMember();
+                      if ( $member &&  ($member == $vitrine->getMembretux()) ) {
+                              $hasAccess = true;
+                      }
+              }
+    }
+    if(! $hasAccess) {
+            throw $this->createAccessDeniedException("You cannot access the requested ressource!");
     }
 
-        return $this->render('vitrine_tux/carte_show.html.twig', [
-            'carte_tux' => $carte,
-            'vitrine_tux' => $vitrine
-        ]);
+    return $this->render('vitrine/carte_show.html.twig', [
+            'carte' => $carte,
+              'vitrine' => $vitrine
+      ]);
+
     }
     #[IsGranted('ROLE_USER')]
     #[Route('/{id}/edit', name: 'app_vitrine_tux_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, VitrineTux $vitrineTux, EntityManagerInterface $entityManager): Response
-    {
+    {   
+        $hasAccess = $this->isGranted('ROLE_ADMIN') ||
+            ($this->getUser()==$vitrineTux->getMembretux()->getUser());
+        if(! $hasAccess) {
+            throw $this->createAccessDeniedException("Vous ne pouvez pas accéder à une vitrine qui n'est pas la vôtre.");
+        }
         $form = $this->createForm(VitrineTuxType::class, $vitrineTux);
         $form->handleRequest($request);
 
@@ -102,7 +153,12 @@ class VitrineTuxController extends AbstractController
     #[IsGranted('ROLE_ADMIN')]
     #[Route('/{id}', name: 'app_vitrine_tux_delete', methods: ['POST'])]
     public function delete(Request $request, VitrineTux $vitrineTux, EntityManagerInterface $entityManager): Response
-    {
+    {   
+        $hasAccess = $this->isGranted('ROLE_ADMIN') ||
+            ($this->getUser()==$vitrineTux->getMembretux()->getUser());
+        if(! $hasAccess) {
+            throw $this->createAccessDeniedException("Vous ne pouvez pas accéder à une vitrine qui n'est pas la vôtre.");
+        }
         if ($this->isCsrfTokenValid('delete'.$vitrineTux->getId(), $request->request->get('_token'))) {
             $entityManager->remove($vitrineTux);
             $entityManager->flush();
